requirements.txt
asgiref==3.8.1
bleach==6.1.0
Django==4.2.13
django-cors-headers==4.3.1
django-filter==24.2
django-redis==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.6
inflection==0.5.1
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
webencodings==0.5.1
wheel==0.45.1
